+++ server/index.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <script src='http://localhost:8080/builds/10489871930474.js'></script>
+    <body>
+        <form id="myForm">
+            <label for="text">Texto:</label>
+            <input type="text" id="text" name="text" value="Texto de exemplo"><br><br>
+    
+            <label for="email">Email:</label>
+            <input type="email" id="email" name="email" value="exemplo@email.com"><br><br>
+    
+            <label for="password">Senha:</label>
+            <input type="password" id="password" name="password" value="minhasenha"><br><br>
+    
+            <label for="checkbox">Checkbox:</label>
+            <input type="checkbox" id="checkbox" name="checkbox" checked><br><br>
+    
+            <label for="radio1">Radio 1:</label>
+            <input type="radio" id="radio1" name="radio" value="opcao1" checked>
+            <label for="radio2">Radio 2:</label>
+            <input type="radio" id="radio2" name="radio" value="opcao2"><br><br>
+    
+            <label for="textarea">Textarea:</label><br>
+            <textarea id="textarea" name="textarea" rows="4" cols="50">Texto de exemplo para textarea.</textarea><br><br>
+    
+            <label for="select">Select:</label>
+            <select id="select" name="select">
+                <option value="opcao1">Opção 1</option>
+                <option value="opcao2" selected>Opção 2</option>
+                <option value="opcao3">Opção 3</option>
+            </select><br><br>
+    
+            <input type="submit" value="Enviar">
+        </form>
+    
+</body>
+</html>
